Refactor: ProductManager에서 ProductList 분리
